tidb support cura

+++ util/chunk/chunk.go
@@ -158,7 +158,7 @@ func newVarLenColumn(cap int, old *Column) *Column {
 		estimatedElemLen = (len(old.data) + len(old.data)/8) / old.length
 	}
 	return &Column{
-		offsets:    make([]int64, 1, cap+1),
+		offsets:    make([]int32, 1, cap+1),
 		data:       make([]byte, 0, cap*estimatedElemLen),
 		nullBitmap: make([]byte, 0, (cap+7)>>3),
 	}
@@ -407,7 +407,7 @@ func appendCellByCell(dst *Column, src *Column, rowIdx int) {
 	} else {
 		start, end := src.offsets[rowIdx], src.offsets[rowIdx+1]
 		dst.data = append(dst.data, src.data[start:end]...)
-		dst.offsets = append(dst.offsets, int64(len(dst.data)))
+		dst.offsets = append(dst.offsets, int32(len(dst.data)))
 	}
 	dst.length++
 }
@@ -430,7 +430,7 @@ func (c *Chunk) preAlloc(row Row) (rowIdx uint32) {
 		elemLen := len(srcCol.elemBuf)
 		if !srcCol.isFixed() {
 			elemLen = int(srcCol.offsets[row.idx+1] - srcCol.offsets[row.idx])
-			dstCol.offsets = append(dstCol.offsets, int64(len(dstCol.data)+elemLen))
+			dstCol.offsets = append(dstCol.offsets, int32(len(dstCol.data)+elemLen))
 		}
 		dstCol.length++
 		needCap := len(dstCol.data) + elemLen
