+++ util/chunk/mutrow.go
@@ -190,7 +190,7 @@ func newMutRowVarLenColumn(valSize int) *Column {
 	buf := make([]byte, valSize+1)
 	col := &Column{
 		length:     1,
-		offsets:    []int64{0, int64(valSize)},
+		offsets:    []int32{0, int32(valSize)},
 		data:       buf[:valSize],
 		nullBitmap: buf[valSize:],
 	}
@@ -325,7 +325,7 @@ func setMutRowBytes(col *Column, bin []byte) {
 		col.nullBitmap = buf[len(bin):]
 	}
 	copy(col.data, bin)
-	col.offsets[1] = int64(len(bin))
+	col.offsets[1] = int32(len(bin))
 }
 
 func setMutRowNameValue(col *Column, name string, val uint64) {
@@ -339,7 +339,7 @@ func setMutRowNameValue(col *Column, name string, val uint64) {
 	}
 	binary.LittleEndian.PutUint64(col.data, val)
 	copy(col.data[8:], name)
-	col.offsets[1] = int64(dataLen)
+	col.offsets[1] = int32(dataLen)
 }
 
 func setMutRowJSON(col *Column, j json.BinaryJSON) {
@@ -355,7 +355,7 @@ func setMutRowJSON(col *Column, j json.BinaryJSON) {
 	}
 	col.data[0] = j.TypeCode
 	copy(col.data[1:], j.Value)
-	col.offsets[1] = int64(dataLen)
+	col.offsets[1] = int32(dataLen)
 }
 
 // ShallowCopyPartialRow shallow copies the data of `row` to MutRow.
@@ -376,7 +376,7 @@ func (mr MutRow) ShallowCopyPartialRow(colIdx int, row Row) {
 		} else {
 			start, end := srcCol.offsets[row.idx], srcCol.offsets[row.idx+1]
 			dstCol.data = srcCol.data[start:end]
-			dstCol.offsets[1] = int64(len(dstCol.data))
+			dstCol.offsets[1] = int32(len(dstCol.data))
 		}
 	}
 }
